@@ -1,33 +1,43 @@
 {
-  "private": true,
-  "scripts": {
-    "dev": "npm run development",
-    "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress  --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "watch": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --progress  --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "watch-poll": "npm run watch -- --watch-poll",
-    "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-    "prod": "npm run production",
-    "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
-  },
-  "devDependencies": {
-    "axios": "^0.27.2",
-    "bootstrap-sass": "^3.3.7",
-    "cross-env": "^5.0.1",
-    "jquery": "^3.1.1",
-    "laravel-mix": "^6.0.49",
-    "lodash": "^4.17.4",
-    "node-sass": "^7.0.1",
-    "resolve-url-loader": "^5.0.0",
-    "safe-buffer": "^5.0.1",
-    "sass": "^1.52.3",
-    "sass-loader": "^12.6.0",
-    "vue": "^2.1.10",
-    "vue-loader": "^15.9.8",
-    "vue-template-compiler": "^2.6.10"
-  },
-  "engines": {
-    "node": ">= 10.18.1",
-    "npm": ">= 6.13.4",
-    "yarn": ">= 1.21.1"
-  }
+    "private": true,
+    "scripts": {
+        "dev": "npm run development",
+        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress  --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "watch": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --progress  --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "watch-poll": "npm run watch -- --watch-poll",
+        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "prod": "npm run production",
+        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+    },
+    "devDependencies": {
+        "axios": "^0.27.2",
+        "bootstrap-sass": "^3.3.7",
+        "core-js": "^3.6.5",
+        "cross-env": "^5.0.1",
+        "jquery": "^3.1.1",
+        "laravel-echo": "^1.11.0",
+        "laravel-mix": "^6.0.49",
+        "lodash": "^4.17.4",
+        "node-sass": "^7.0.1",
+        "pusher-js": "^7.0.3",
+        "resolve-url-loader": "^5.0.0",
+        "safe-buffer": "^5.0.1",
+        "sass": "^1.52.3",
+        "sass-loader": "^12.6.0",
+        "timeago.js": "^4.0.2",
+        "vue-i18n": "^9.1.6",
+        "vue-loader": "^16.1.2",
+        "vue-router": "^4.0.0",
+        "vue-template-compiler": "^2.6.10",
+        "vuex": "^4.0.1"
+    },
+    "engines": {
+        "node": ">= 10.18.1",
+        "npm": ">= 6.13.4",
+        "yarn": ">= 1.21.1"
+    },
+    "dependencies": {
+        "quasar": "^2.7.1",
+        "vue": "^3.2.37"
+    }
 }
