@@ -1,4 +1,7 @@
-let mix = require('laravel-mix');
+const mix = require('laravel-mix')
 
-mix.js('resources/assets/js/app.js', 'public/js').vue();
-   // .sass('resources/assets/sass/app.scss', 'public/css');
+// mix.js('resources/assets/js/app.js', 'public/js').vue()
+
+mix
+    .copy('public/spa/index.html', 'resources/views/converses/login.blade.php')
+    .copyDirectory('public/spa', 'public')
